Cleanup
Also added option to disable moss spread from water
Also added light level check for moss spread, didn't feel right to have
it grow in full sunlight
Made spores decrease the stack size when planting mushrooms

+++ src/main/java/exnihiloomnia/items/buckets/ItemBucketPorcelain.java
@@ -65,19 +65,19 @@ public class ItemBucketPorcelain extends ItemBucket {
 					IBlockState iblockstate = worldIn.getBlockState(blockpos);
 					Material material = iblockstate.getMaterial();
 
-					if (worldIn.getBlockState(blockpos).getBlock().equals(ENOBlocks.WITCHWATER) && (iblockstate.getValue(BlockLiquid.LEVEL)).intValue() == 0) {
+					if (worldIn.getBlockState(blockpos).getBlock().equals(ENOBlocks.WITCHWATER) && iblockstate.getValue(BlockLiquid.LEVEL) == 0) {
 						playerIn.playSound(SoundEvents.ITEM_BUCKET_FILL, 1.0F, 1.0F);
 						worldIn.setBlockState(blockpos, Blocks.AIR.getDefaultState(), 11);
 						playerIn.addStat(StatList.getObjectUseStats(this));
 						return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, this.fillBucket(itemStackIn, playerIn, ENOItems.BUCKET_PORCELAIN_WITCHWATER));
 					}
-					else if (material == Material.WATER && (iblockstate.getValue(BlockLiquid.LEVEL)).intValue() == 0) {
+					else if (material == Material.WATER && iblockstate.getValue(BlockLiquid.LEVEL) == 0) {
 						worldIn.setBlockState(blockpos, Blocks.AIR.getDefaultState(), 11);
 						playerIn.addStat(StatList.getObjectUseStats(this));
 						playerIn.playSound(SoundEvents.ITEM_BUCKET_FILL, 1.0F, 1.0F);
 						return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, this.fillBucket(itemStackIn, playerIn, ENOItems.BUCKET_PORCELAIN_WATER));
 					}
-					else if (material == Material.LAVA && (iblockstate.getValue(BlockLiquid.LEVEL)).intValue() == 0) {
+					else if (material == Material.LAVA && iblockstate.getValue(BlockLiquid.LEVEL) == 0) {
 						playerIn.playSound(SoundEvents.ITEM_BUCKET_FILL_LAVA, 1.0F, 1.0F);
 						worldIn.setBlockState(blockpos, Blocks.AIR.getDefaultState(), 11);
 						playerIn.addStat(StatList.getObjectUseStats(this));
